RPC Discover endpoint with basic informations
