+++ crates/bevy_remote/src/schemas/mod.rs
@@ -1,0 +1,4 @@
+//! Module with schemas used for various BRP endpoints
+
+pub mod json_schema;
+pub mod open_rpc;
